split menu components
