+++ src/components/Layout/Header/Menu/LoginMenuItem.js
@@ -1,0 +1,12 @@
+import React, { useContext } from 'react';
+import AuthContext from '../../../../context/Authentication';
+
+import MenuItemParent from './MenuItemParent';
+import MenuItem from './MenuItem';
+
+const LoginMenuItem = () => {
+  const { customUserData } = useContext(AuthContext);
+  const signOut = useSignOut();
+
+  return <MenuItemParent />;
+};
