+++ src/components/Layout/Header/index.js
@@ -1,11 +1,13 @@
 import React, { useState } from 'react';
-import s from './style.module.less';
 import Link from 'gatsby-link';
-import Logo from './logo.svg';
-import cN from 'classnames';
+
 import { OverlayContext } from '../../../context/Overlay';
 import { Button } from '../../Forms/Button';
 // import { Tooltip } from '../../Tooltip';
+
+import s from './style.module.less';
+import Logo from './logo.svg';
+import Menu from './Menu';
 
 const Header = ({ menu, hasOverlay }) => {
   const [menuOpen, setMenuOpen] = useState(false);
@@ -71,75 +73,5 @@ const Header = ({ menu, hasOverlay }) => {
     </header>
   );
 };
-
-const Menu = ({ menu, menuOpen }) => {
-  return (
-    <ul className={cN(s.navList, { [s.isOpen]: menuOpen })} id="menuHeader">
-      {menu.map((item, index) => {
-        if (item.__typename === 'ContentfulStaticContent') {
-          return <MenuItem key={index} {...item} />;
-        } else {
-          return <MenuItemParent key={index} {...item} />;
-        }
-      })}
-    </ul>
-  );
-};
-
-const MenuItemParent = ({
-  title,
-  contentfulchildren,
-  internalLink,
-  externalLink,
-}) => {
-  const children = contentfulchildren;
-  return (
-    <li className={cN(s.navItem, s.navItemParent)}>
-      {!internalLink && !externalLink && (
-        <div className={s.menuItemParentTitle}>{title}</div>
-      )}
-      {internalLink && (
-        <Link
-          className={s.link}
-          activeClassName={s.linkActive}
-          to={internalLink}
-        >
-          {title}
-        </Link>
-      )}
-      {externalLink && (
-        <a
-          className={s.link}
-          href={externalLink}
-          target="_blank"
-          rel="noreferrer"
-        >
-          {title}
-        </a>
-      )}
-      {children && (
-        <div className={s.menuItemParentChildren}>
-          <ul className={s.menuItemParentChildrenInner}>
-            {children.map((item, index) => (
-              <MenuItem key={index} child={true} {...item} />
-            ))}
-          </ul>
-        </div>
-      )}
-    </li>
-  );
-};
-
-const MenuItem = ({ slug, shortTitle, title, child }) => (
-  <li className={cN(s.navItem, { [s.navItemChild]: child })}>
-    <Link
-      className={s.link}
-      activeClassName={s.linkActive}
-      to={slug === '/' ? '/' : `/${slug}/`}
-    >
-      {shortTitle || title}
-    </Link>
-  </li>
-);
 
 export default Header;
